voucher

+++ src/main/java/com/ecom/service/VoucherService.java
@@ -7,19 +7,19 @@ import com.ecom.model.Voucher;
 public interface VoucherService {
 
     // Create a new voucher for a user
-    Voucher createVoucher(Long userId, Integer discountPercent);
+    Voucher createVoucher(Integer userId, Integer discountPercent);
 
     // Get valid vouchers for a user (not used and not expired)
-    List<Voucher> getUserValidVouchers(Long userId);
+    List<Voucher> getUserValidVouchers(Integer userId);
 
     // Get all vouchers for a user
-    List<Voucher> getAllUserVouchers(Long userId);
+    List<Voucher> getAllUserVouchers(Integer userId);
 
     // Get voucher by code
     Voucher getVoucherByCode(String code);
 
     // Apply voucher (mark as used)
-    Boolean applyVoucher(String code, Long userId);
+    Boolean applyVoucher(String code, Integer userId);
 
     // Generate unique voucher code
     String generateVoucherCode();
